feat: api-handlers-decomposition-recovery-005 - Extract events and status route handlers
