+++ docs/internal/processes/development-guide-relevant-files.md
@@ -141,7 +141,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/api/handlers_factory.go` | Factory validation route handlers | Owns `POST /factory/validate` (`ValidateFactory`), `decodeNamedFactoryBody`, and factory-definition validation response mapping via `factoryvalidation.Validate` / `ToValidationTargets`. |
 | `pkg/api/handlers_factory_sessions.go` | Factory session and current-factory route handlers | Owns `/factory-sessions` list/open/close, default and session-scoped current-factory GET, `PUT /factory-sessions/{session_id}/factory` (`SaveCurrentFactoryBySessionId` with `{ mode?, factory }` → `SaveFactoryForSession` only), workstation prompt-template contract/validate routes, `loadCurrentFactory*` / `writeCurrentFactoryError`, and session/factory body decode helpers. |
 | `pkg/api/handlers_models.go` | Model route handlers | Owns `GET /models`, `GET /models/{model_name}`, `POST /models/{model_name}/invoke`, `POST /models/{model_name}/pull`, plus `decodeModelInvocationRequestBody` and model-invocation response projection helpers. |
-| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep events/status routes and shared token/status projection helpers in `handlers.go` while route-family handlers live in `handlers_work.go`, `handlers_factory.go`, `handlers_factory_sessions.go`, `handlers_models.go`, and follow-on `handlers_events.go`; shared JSON/error writers live in `handlers_common.go`. Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
+| `pkg/api/handlers_*.go` files and `pkg/api/server.go` | Handwritten API boundary | Route-family handlers live in `handlers_work.go`, `handlers_factory.go`, `handlers_factory_sessions.go`, `handlers_models.go`, and `handlers_events.go` (`GetStatus`, `GetEvents`, SSE streaming, `tokenToResponse`, status token categorization); shared JSON/error writers and cross-route optional-field helpers live in `handlers_common.go`; `server.go` owns `ServerInterface` compile assertion. Keep `pkg/api` at the 15-file cap by deleting legacy `handlers.go` when adding `handlers_events.go`. Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
 | `pkg/api/server_*_test.go` plus `pkg/api/server_test_helpers_test.go` | API route regression ownership | When API route tests cross the backend size gate, split them into same-package sibling `server_*_test.go` files by boundary concern such as submit, upsert, factory, provider-session, work-query, and dashboard/event coverage, and keep shared request builders or response assertions in `server_test_helpers_test.go` so route behavior stays observable without recreating one giant test owner file. |
 | `api/codegen_config/client.yaml` | Generated Go client config | The checked-in Go client for backend functional tests is generated from this config and must stay in lockstep with `api/openapi.yaml`, `pkg/generatedclient/`, and `make api-smoke`. |
 | `api/openapi-main.yaml` | Authored OpenAPI source of truth | Public API behavior changes should start here, including the advertised local `servers` URL, then refresh the bundled spec and generated Go/UI artifacts together so handwritten route seams and generated parameter binding stay aligned. |
